third_party/blink/web_tests/crypto/subtle/aes-kw/wrapKey-unwrapKey-raw-expected.txt

Tests wrapping and unwrapping of AES-CBC keys using AES-KW and raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS: Wrapped key data should be [1fa68b0a8112b447aef34bd8fb5a7b829d3e862371d2cfe5] and was
PASS unwrappedKey.algorithm.name is "HMAC"
PASS unwrappedKey.algorithm.hash.name is "SHA-1"
PASS unwrappedKey.algorithm.length is 128
PASS typeof unwrappedKey.extractable is 'boolean'
PASS unwrappedKey.extractable is true
PASS unwrappedKey.usages.join(',') is "sign"
PASS: Unwrapped key data should be [00112233445566778899aabbccddeeff] and was
PASS: Wrapped key data should be [64e8c3f9ce0f5ba263e9777905818a2a93c8191e7d6e8ae7] and was
PASS unwrappedKey.algorithm.name is "HMAC"
PASS unwrappedKey.algorithm.hash.name is "SHA-1"
PASS unwrappedKey.algorithm.length is 128
PASS typeof unwrappedKey.extractable is 'boolean'
PASS unwrappedKey.extractable is true
PASS unwrappedKey.usages.join(',') is "sign"
PASS: Unwrapped key data should be [00112233445566778899aabbccddeeff] and was
PASS: Wrapped key data should be [a8f9bc1612c68b3ff6e6f4fbe30e71e4769c8b80a32cb8958cd5d17d6b254da1] and was
PASS unwrappedKey.algorithm.name is "HMAC"
PASS unwrappedKey.algorithm.hash.name is "SHA-1"
PASS unwrappedKey.algorithm.length is 192
PASS typeof unwrappedKey.extractable is 'boolean'
PASS unwrappedKey.extractable is true
PASS unwrappedKey.usages.join(',') is "sign"
PASS: Unwrapped key data should be [00112233445566778899aabbccddeeff0001020304050607] and was
PASS: Wrapped key data should be [28c9f404c4b810f4cbccb35cfb87f8263f5786e2d80ed326cbc7f0e71a99f43bfb988b9b7a02dd21] and was
PASS unwrappedKey.algorithm.name is "HMAC"
PASS unwrappedKey.algorithm.hash.name is "SHA-1"
PASS unwrappedKey.algorithm.length is 256
PASS typeof unwrappedKey.extractable is 'boolean'
PASS unwrappedKey.extractable is true
PASS unwrappedKey.usages.join(',') is "sign"
PASS: Unwrapped key data should be [00112233445566778899aabbccddeeff000102030405060708090a0b0c0d0e0f] and was
PASS successfullyParsed is true

TEST COMPLETE
